translation: extract RuleCompiler, wire Exporter.translations into ir.json
New ADR-0030-in-progress step 1. Extracted the PURE half of head_
compiler.rb's SQL compiler (compile_rules/compile_compute/compile_id_
expression/id_case/rekeyed? -- no database connection, no watermark,
no era chain) into Translation::RuleCompiler, a standalone module with
no adapter coupling. head_compiler.rb's own per-mint assembly now
calls into it instead of defining these methods itself -- same code,
zero duplication, verified byte-identical behavior against real
Postgres (45 examples, spec/adapters/driven/postgres_era/).

This is what lets Exporter.translation_aggregate carry each edge's
precompiled SQL (compiled_state_expression/compiled_id_expression)
without Exporter depending on the PostgresEra adapter directly --
head_compiler.rb's real per-mint assembly and this build-time export
now call the exact same compiler, so a future Rust-side mint executor
gets Ruby's own compiler's output verbatim, never a second,
independently-authored one that could drift.

Wired into bin/project_rust: target_registry now loads bluebook/
translations/*.bluebook (it never did before -- Exporter.translations
had nothing real to export from this generator's own registry build),
and target_ir[:translations] rides alongside the existing [:lineage]
key into ir.json, the same sidecar rust/host already reads generically
at runtime via HECKS_IR_PATH.

Opened a new, honestly-scoped exception in spec/project_rust_pipeline_
spec.rb: ir.json and metadata.rs now differ between the default Ruby
path and the opt-in Rust-native pipeline (which has no translation_
pass.rs counterpart to lineage_pass.rs yet) -- whole-file skip for
both, not a surgical key-strip, since metadata.rs carries the same
JSON escaped through Rust's own string-literal rules. Documented what
would close it.

Verified against real translation edges, not synthetic ones: examples/
pizzas' own real Pizza->Order rename+move edge now produces the exact
head_compiler.rb-equivalent SQL in ir.json. Full suite (1655 examples),
the real-Postgres lineage suite (45 examples), rust_conformance_spec.rb
(16 examples), and project_rust_pipeline_spec.rb (2 examples, both
pizzas and banking) all green.
