+++ lib/hecksagain/translation.rb
@@ -5,6 +5,7 @@ module Hecksagain
   end
 end
 
+require_relative "translation/rule_compiler"
 require_relative "translation/scaffold"
 require_relative "translation/audit"
 require_relative "translation/reattest"
